// packages/rath-client/src/interfaces.ts
export interface IRow {
    [key: string]: any
}

export interface ICol<T> {
    fid: string;
    data: Array<T>;
}

export type IGeoRole = 'longitude' | 'latitude' | 'none';

export type FieldExtOptType = (
    | 'dateTimeExpand'
    | `CC.${string}`
);

/** Detailed information of a extended field.  */
interface IFieldExtInfoBase {
    /** Field id of fields that this field inferred from.  */
    extFrom: string[];
    /** The identifier of the data-extension operation. */
    extOpt: FieldExtOptType;
    /** Additional information of the specified extension operation */
    extInfo: any;
}
// ...
export interface IMuteFieldBase {
    fid: string;
    name?: string;
    analyticType: 'dimension' | 'measure' | '?';
    semanticType: 'nominal' | 'temporal' | 'ordinal' | 'quantitative' | '?';
    disable?: boolean | '?';
    geoRole: IGeoRole | '?';
    /** detailed information of field extension operations. Non-null if this field is extended */
    extInfo?: IFieldExtInfoBase;
}
// ...
export enum IDataPreviewMode {
    meta = 'meta',
    data = 'data',
    stat = 'stat'
}

export enum ITaskTestMode {
    local = 'local',
    server = 'server'
}

export enum IResizeMode {
    auto = 'auto',
    control = 'control'
}

export type { IFilter }

export type CleanMethod = 'dropNull' | 'useMode' | 'simpleClean' | 'none';

export enum IDataPrepProgressTag {
    none = 'none',
    upload = 'upload',
    filter = 'filter',
    clean = 'clean'
}
// ...
export interface IDataFetchRecipe {
    type: 'database' | 'restful';
    // Database
    dbServer?: string;
    dbSourceType?: string;
    dbUri?: string;
    dbQuery?: string;
    dbCredentials?: Record<string, string>;
    // REST / JSON API
    apiUrl?: string;
    apiFormat?: string;
}